References #2, added initial implementation.
The page is overridden for both authenticated and unauthenticated users,
that would probably change later. The texts could be entered from the
YCHANGE Plugin Settings page (these do not need their own URLs nor need
to be publically accessible). The choice of using Settings is
questionable, but it should do not the starters. Extending externalpages
plugin did not seem to be a meaningful process and would have required
almost a total rewrite of the codebase.